Fix bug bioserver not found and move BioServerCommand to Command API

+++ ui.new/backend/ServerCommand.cpp
@@ -11,15 +11,23 @@ ServerCommand::~ServerCommand() {
 }
 
 void ServerCommand::terminate() {
-    _process->terminate();
-    _process->wait();
+    if (_process) {
+        _process->terminate();
+        _process->wait();
+    }
 }
 
 bool ServerCommand::isDone() {
+    if (!_process) {
+        return true;
+    }
     return !_process->running();
 }
 
 int ServerCommand::getReturnCode() const {
+    if (!_process) {
+        return 1;
+    }
     return _process->exit_code();
 }
